fix: #1639 services/ 層 DB 直接 import を解消 (アーキテクチャテスト対応)
unit-test (2) の tests/unit/arch/no-direct-db-access.test.ts が
`analytics-service.ts` から `$lib/server/db/dynamodb/client` への直接 import を検出して FAIL。

修正:
- DynamoDB GSI2 query helper (`queryAnalyticsEventTenants`) を
  `$lib/analytics/providers/dynamo.ts` に移動
- `analytics-service.ts` は `queryAnalyticsEventTenants` を import するだけにし、
  services/ 層の DB 直接アクセス禁止 (#1021) に準拠
- 単体テスト: queryAnalyticsEventTenants の mock に置き換え (3 ケース → 既存 23 全通過)
- afterEach に vi.doUnmock を追加し、後続 describe (AnalyticsManager) で module mock が
  残らないようガード

cspell も「granularities」を「period granularity」に変更して通過。

closes #1639 (CI fix)
